depot: fix locale influence on hash generation
We used to export 'LC_COLLATE=C' for the environment of depot tools. This
ensures, for instance, that 'depot/extract' always uses the same 'sort' style
in the code that generates the hash of extracted content. This is required,
so, that the same content generates the same hash in different environments.
However, 'LC_COLLATE' might not be sufficient. If a user has set 'LC_ALL'
non-empty and to something different than 'C' when calling 'depot/extract' the
'sort' will use the user locale instead of 'C'.

The manual of 'sort' recommends to set 'LC_ALL=C' if you want reliable
results, so, we do this now locally for the affected 'sort' calls.

Fixes #4144